+++ flake.nix
@@ -9,7 +9,7 @@
   };
   outputs = { self, nixpkgs }:
     let
-      annealVersion = "0.24.1";
+      annealVersion = "0.25.0";
       systems = [ "aarch64-darwin" "x86_64-darwin" "aarch64-linux" "x86_64-linux" ];
       forAllSystems = f: nixpkgs.lib.genAttrs systems (system: f {
         pkgs = nixpkgs.legacyPackages.${system};